+++ src/vfxnaming/naming.py
@@ -24,7 +24,7 @@ import shutil
 import vfxnaming.rules as rules
 import vfxnaming.tokens as tokens
 from pathlib import Path
-from typing import AnyStr, Dict, Union
+from typing import AnyStr, Dict, Union, Iterable
 
 from vfxnaming.logger import logger
 from vfxnaming.error import SolvingError, RepoError
@@ -138,9 +138,12 @@ def solve(*args, **kwargs) -> AnyStr:
     return rule.solve(**values)
 
 
-def validate(name: AnyStr, **kwargs) -> bool:
-    """Validates a name string against the currently active rule and its
-    tokens if passed as keyword arguments.
+def validate(  # noqa: C901
+    name: AnyStr, with_rules: Iterable[str] = [], strict: bool = False, **kwargs
+) -> Iterable[rules.Rule]:
+    """Validates a name string against the currently active rule if no rules are passed or
+    against the list of specific rules passed in with_rules.
+    It also validates its tokens if passed as keyword arguments.
 
     -For rules with repeated tokens:
 
@@ -162,50 +165,77 @@ def validate(name: AnyStr, **kwargs) -> bool:
     Args:
         name (str): Name string e.g.: C_helmet_001_MSH
 
+        with_rules (list, optional): List of rule names to validate against. Defaults to [].
+
+        strict (bool, optional): If False, it'll try to accept casing mismatches.
+
         kwargs (dict): Keyword arguments with token names and values.
 
     Returns:
-        bool: True if the name is valid, False otherwise.
+        list: List of validated rules. Empty list if no rule could be validated.
     """
-    rule = rules.get_active_rule()
-    # * This accounts for those cases where a token is used more than once in a rule
-    repeated_fields = dict()
-    for each in rule.fields:
-        if each not in repeated_fields.keys():
-            if rule.fields.count(each) > 1:
-                repeated_fields[each] = 1
-    fields_with_digits = list()
-    for each in rule.fields:
-        if each in repeated_fields.keys():
-            counter = repeated_fields.get(each)
-            repeated_fields[each] = counter + 1
-            fields_with_digits.append(f"{each}{counter}")
-        else:
-            fields_with_digits.append(each)
-    values = {}
-    fields_inc = 0
-    for f in fields_with_digits:
-        token = tokens.get_token(rule.fields[fields_inc])
-        if token:
-            # Explicitly passed as keyword argument
-            if kwargs.get(f) is not None:
-                values[f] = token.solve(kwargs.get(f))
-                fields_inc += 1
-                continue
-            # Explicitly passed as keyword argument without repetitive digits
-            # Use passed argument for all field repetitions
-            elif kwargs.get(rule.fields[fields_inc]) is not None:
-                values[f] = token.solve(kwargs.get(rule.fields[fields_inc]))
-                fields_inc += 1
-                continue
-            elif token.required and isinstance(token, tokens.Token):
-                if len(token.fallback):
-                    values[f] = token.fallback
+    previously_active_rule = rules.get_active_rule()
+    if not len(with_rules):
+        with_rules = [previously_active_rule.name]
+    validated: Iterable[rules.Rule] = []
+    for with_rule in with_rules:
+        rule = rules.get_rule(with_rule)
+        if not rule:
+            logger.warning(f"Rule {with_rule} not found.")
+
+        rules.set_active_rule(rule)
+        # * This accounts for those cases where a token is used more than once in a rule
+        repeated_fields = dict()
+        for each in rule.fields:
+            if each not in repeated_fields.keys():
+                if rule.fields.count(each) > 1:
+                    repeated_fields[each] = 1
+        fields_with_digits = list()
+        for each in rule.fields:
+            if each in repeated_fields.keys():
+                counter = repeated_fields.get(each)
+                repeated_fields[each] = counter + 1
+                fields_with_digits.append(f"{each}{counter}")
+            else:
+                fields_with_digits.append(each)
+        values = {}
+        fields_inc = 0
+        for f in fields_with_digits:
+            token = tokens.get_token(rule.fields[fields_inc])
+            if token:
+                # Explicitly passed as keyword argument
+                if kwargs.get(f) is not None:
+                    values[f] = token.solve(kwargs.get(f))
                     fields_inc += 1
                     continue
-            fields_inc += 1
-    logger.debug(f"Validating rule '{rule.name}' with values {values}")
-    return rule.validate(name, **values)
+                # Explicitly passed as keyword argument without repetitive digits
+                # Use passed argument for all field repetitions
+                elif kwargs.get(rule.fields[fields_inc]) is not None:
+                    values[f] = token.solve(kwargs.get(rule.fields[fields_inc]))
+                    fields_inc += 1
+                    continue
+                elif token.required and isinstance(token, tokens.Token):
+                    if len(token.fallback):
+                        values[f] = token.fallback
+                        fields_inc += 1
+                        continue
+                fields_inc += 1
+        logger.debug(f"Validating rule '{rule.name}' with values {values}")
+        validation = rule.validate(name, strict, **values)
+        if validation:
+            rules.set_active_rule(previously_active_rule)
+            validated.append(rule)
+    rules.set_active_rule(previously_active_rule)
+    if not len(validated):
+        logger.warning(
+            f"Could not validate {name} with any of the given "
+            f"rules {', '.join([rule for rule in with_rules])}."
+        )
+    else:
+        logger.info(
+            f"Name {name} validated with rules: {', '.join([rule.name for rule in validated])}."
+        )
+    return validated
 
 
 def validate_repo(repo: Path) -> bool: